agent/requirements.txt
fastapi
uvicorn
requests
firebase-admin
google-adk
